requirements.txt
requests
pillow
typing-extensions
lxml
beautifulsoup4
python-dotenv
pymongo
